backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.2.0
python-dotenv==1.0.0
httpx==0.26.0
email-validator==2.3.0
reportlab==4.4.10
python-docx==1.2.0
PyPDF2==3.0.1
authlib==1.6.8
PyMuPDF==1.27.1
opencv-python==4.12.0.88
ezdxf==1.4.3
numpy==2.2.6

